Improve plugin loading
