--- bff/testdata/plugins-error.yaml
@@ -1,2 +1,0 @@
-plugins:
-  - name: [1, 2, 3]
