+++ bff/transport/rest/misc/config.go
@@ -9,6 +9,6 @@ func Config(c *gin.Context) {
 	c.JSON(200, gin.H{
 		"emailAddresses": config.EMAIL_ADDRESSES,
 		"disableProxy":   !config.PROXY_ENABLE,
-		"plugins":        config.PLUGINS,
+		"plugins":        config.Plugins,
 	})
 }
